requirements.txt
transformers
transformers[sentencepiece]
datasets
sacrebleu
rouge_score
pandas
py7zr
nltk
tqdm
PyYAML
matplotlib
torch
notebook
boto3
mypy-boto3-s3
python-box==6.0.2
ensure==1.0.2
fastapi==0.78.0
uvicorn==0.18.3
Jinja2==3.1.2
-e .
 